Move updatable template types to /models
