+++ src/functions/templates/update-one-template.spec.ts
@@ -7,14 +7,14 @@
  **************************************************************************/
 
 import { omit } from 'lodash';
-import { CreatableTemplate, isTemplate, Template } from '../../models';
+import { CreatableTemplate, isTemplate, Template, UpdatableTemplate } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { omitUndefinedShallow } from '../utils';
 import { makeCreateOneTemplate } from './create-one-template';
 import { makeDeleteOneTemplate } from './delete-one-template';
 import { makeGetOneTemplate } from './get-one-template';
-import { makeUpdateOneTemplate, UpdatableTemplate } from './update-one-template';
+import { makeUpdateOneTemplate } from './update-one-template';
 
 describe('updateOneTemplate()', () => {
 	const createOneTemplate = makeCreateOneTemplate({ host: TEST_HOST, useEncryption: false });
