+++ src/functions/templates/update-one-template.ts
@@ -6,9 +6,7 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isUndefined } from 'lodash';
-import { RawTemplate, Template, toTemplate } from '../../models';
-import { NumericID, RawNumericID, toRawNumericID, UUID } from '../../value-objects';
+import { RawTemplate, Template, toRawUpdatableTemplate, toTemplate, UpdatableTemplate } from '../../models';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
@@ -46,72 +44,5 @@ export const makeUpdateOneTemplate = (makerOptions: APIFunctionMakerOptions) => 
 			if (err instanceof Error) throw err;
 			throw Error('Unknown error');
 		}
-	};
-};
-
-export interface UpdatableTemplate {
-	uuid: UUID;
-
-	userID?: NumericID;
-	groupIDs?: Array<NumericID>;
-
-	name?: string;
-	description?: string | null;
-	labels?: Array<string>;
-
-	isGlobal?: boolean;
-	isRequired?: boolean;
-
-	query?: string;
-	variable?: {
-		token?: string;
-		name?: string;
-		description?: string | null;
-	};
-
-	previewValue?: string | null;
-}
-
-interface RawUpdatableTemplate {
-	UID: RawNumericID;
-	GIDs: Array<RawNumericID>;
-
-	Global: boolean;
-	Labels: Array<string>;
-
-	Name: string;
-	Description: string | null; // Null becomes empty string
-	Contents: {
-		query: string;
-		variable: string;
-		variableLabel: string;
-		variableDescription: string | null;
-		required: boolean;
-		testValue: string | null;
-	};
-}
-
-const toRawUpdatableTemplate = (updatable: UpdatableTemplate, current: Template): RawUpdatableTemplate => {
-	const updatableVariableDescription = updatable.variable?.description;
-
-	return {
-		UID: toRawNumericID(updatable.userID ?? current.userID),
-		GIDs: (updatable.groupIDs ?? current.groupIDs).map(toRawNumericID),
-
-		Global: updatable.isGlobal ?? current.isGlobal,
-		Labels: updatable.labels ?? current.labels,
-
-		Name: updatable.name ?? current.name,
-		Description: isUndefined(updatable.description) ? current.description : updatable.description,
-		Contents: {
-			required: updatable.isRequired ?? current.isRequired,
-			query: updatable.query ?? current.query,
-			variable: updatable.variable?.token ?? current.variable.token,
-			variableLabel: updatable.variable?.name ?? current.variable.name,
-			variableDescription: isUndefined(updatableVariableDescription)
-				? current.variable.description
-				: updatableVariableDescription,
-			testValue: isUndefined(updatable.previewValue) ? current.previewValue : updatable.previewValue,
-		},
 	};
 };
