+++ src/models/template/index.ts
@@ -10,6 +10,9 @@ export * from './creatable-template';
 export * from './is-template';
 export * from './raw-creatable-template';
 export * from './raw-template';
+export * from './raw-updatable-template';
 export * from './template';
 export * from './to-raw-creatable-template';
+export * from './to-raw-updatable-template';
 export * from './to-template';
+export * from './updatable-template';
