+++ src/models/template/to-raw-updatable-template.ts
@@ -1,0 +1,38 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { isUndefined } from 'lodash';
+import { toRawNumericID } from '../../value-objects';
+import { RawUpdatableTemplate } from './raw-updatable-template';
+import { Template } from './template';
+import { UpdatableTemplate } from './updatable-template';
+
+export const toRawUpdatableTemplate = (updatable: UpdatableTemplate, current: Template): RawUpdatableTemplate => {
+	const updatableVariableDescription = updatable.variable?.description;
+
+	return {
+		UID: toRawNumericID(updatable.userID ?? current.userID),
+		GIDs: (updatable.groupIDs ?? current.groupIDs).map(toRawNumericID),
+
+		Global: updatable.isGlobal ?? current.isGlobal,
+		Labels: updatable.labels ?? current.labels,
+
+		Name: updatable.name ?? current.name,
+		Description: isUndefined(updatable.description) ? current.description : updatable.description,
+		Contents: {
+			required: updatable.isRequired ?? current.isRequired,
+			query: updatable.query ?? current.query,
+			variable: updatable.variable?.token ?? current.variable.token,
+			variableLabel: updatable.variable?.name ?? current.variable.name,
+			variableDescription: isUndefined(updatableVariableDescription)
+				? current.variable.description
+				: updatableVariableDescription,
+			testValue: isUndefined(updatable.previewValue) ? current.previewValue : updatable.previewValue,
+		},
+	};
+};
